requirements.txt
Django==5.1.3
djangorestframework==3.15.2
gunicorn==21.2.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
python-decouple==3.8